#59 [MOD] ScheduleEntity 필드 type 수정

+++ domain/src/main/java/com/sopt/domain/entity/ScheduleEntity.kt
@@ -15,5 +15,5 @@ data class ScheduleListDetailEntity(
     val category: String = "",
     val startTime: String = "", // H:m - ex) 1:10, 13:01
     val endTime: String = "", // H:m - ex) 1:10, 13:01
-    val duration: Int = 0
+    val duration: Long = 0
 ) : Parcelable
